+++ _stbt/core.py
@@ -281,50 +281,6 @@ def load_image(filename, flags=cv2.IMREAD_COLOR):
     return image
 
 
-class MotionResult(object):
-    """The result from `detect_motion` and `wait_for_motion`.
-
-    :ivar float time: The time at which the video-frame was captured, in
-        seconds since 1970-01-01T00:00Z. This timestamp can be compared with
-        system time (``time.time()``).
-
-    :ivar bool motion: True if motion was found. This is the same as evaluating
-        ``MotionResult`` as a bool. That is, ``if result:`` will behave the
-        same as ``if result.motion:``.
-
-    :ivar Region region: Bounding box where the motion was found, or ``None``
-        if no motion was found.
-
-    :ivar Frame frame: The video frame in which motion was (or wasn't) found.
-
-    :ivar int timestamp: DEPRECATED. Timestamp in nanoseconds. Use ``time``
-        instead.
-
-    Added in v28: The ``frame`` attribute.
-    """
-    def __init__(self, time, motion, region, frame):
-        self.time = time
-        self.motion = motion
-        self.region = region
-        self.frame = frame
-
-    def __nonzero__(self):
-        return self.motion
-
-    def __repr__(self):
-        return (
-            "MotionResult(time=%s, motion=%r, region=%r, frame=%s)" % (
-                "None" if self.time is None else "%.3f" % self.time,
-                self.motion, self.region, _frame_repr(self.frame)))
-
-    @property
-    def timestamp(self):
-        if self.time is None:
-            return None
-        else:
-            return int(self.time * 1e9)
-
-
 class _IsScreenBlackResult(object):
     def __init__(self, black, frame):
         self.black = black
@@ -478,47 +434,6 @@ def new_device_under_test_from_config(
         display=display[0], control=uri_to_control(args.control, display[0]),
         sink_pipeline=sink_pipeline, mainloop=mainloop,
         use_old_threading_behaviour=use_old_threading_behaviour)
-
-
-def _pixel_bounding_box(img):
-    """
-    Find the smallest region that contains all the non-zero pixels in an image.
-
-    >>> _pixel_bounding_box(numpy.array([[0]], dtype=numpy.uint8))
-    >>> _pixel_bounding_box(numpy.array([[1]], dtype=numpy.uint8))
-    Region(x=0, y=0, right=1, bottom=1)
-    >>> _pixel_bounding_box(numpy.array([
-    ...     [0, 0, 0, 0],
-    ...     [0, 1, 1, 1],
-    ...     [0, 1, 1, 1],
-    ...     [0, 0, 0, 0],
-    ... ], dtype=numpy.uint8))
-    Region(x=1, y=1, right=4, bottom=3)
-    >>> _pixel_bounding_box(numpy.array([
-    ...     [0, 0, 0, 0, 0, 0],
-    ...     [0, 0, 0, 1, 0, 0],
-    ...     [0, 1, 0, 0, 0, 0],
-    ...     [0, 0, 0, 0, 1, 0],
-    ...     [0, 0, 1, 0, 0, 0],
-    ...     [0, 0, 0, 0, 0, 0]
-    ... ], dtype=numpy.uint8))
-    Region(x=1, y=1, right=5, bottom=5)
-    """
-    if len(img.shape) != 2:
-        raise ValueError("Single-channel image required.  Provided image has "
-                         "shape %r" % (img.shape,))
-
-    out = [None, None, None, None]
-
-    for axis in (0, 1):
-        flat = numpy.any(img, axis=axis)
-        indices = numpy.where(flat)[0]
-        if len(indices) == 0:
-            return None
-        out[axis] = indices[0]
-        out[axis + 2] = indices[-1] + 1
-
-    return Region.from_extents(*out)
 
 
 class DeviceUnderTest(object):
@@ -698,76 +613,6 @@ class DeviceUnderTest(object):
                     os.path.basename(template.friendly_name))
             yield result
 
-    def detect_motion(self, timeout_secs=10, noise_threshold=None, mask=None,
-                      region=Region.ALL):
-
-        if noise_threshold is None:
-            noise_threshold = get_config(
-                'motion', 'noise_threshold', type_=float)
-
-        debug("Searching for motion")
-
-        if mask is None:
-            mask = _ImageFromUser(None, None, None)
-        else:
-            mask = _load_image(mask, cv2.IMREAD_GRAYSCALE)
-            debug("Using mask %s" % mask.friendly_name)
-
-        frames = self.frames(timeout_secs)
-        frame, _ = next(frames)
-
-        region = Region.intersect(_image_region(frame), region)
-
-        previous_frame_gray = cv2.cvtColor(crop(frame, region),
-                                           cv2.COLOR_BGR2GRAY)
-        if (mask.image is not None and
-                mask.image.shape[:2] != previous_frame_gray.shape[:2]):
-            raise ValueError(
-                "The dimensions of the mask '%s' %s don't match the "
-                "video frame %s" % (
-                    mask.friendly_name, mask.image.shape,
-                    previous_frame_gray.shape))
-
-        for frame, _ in frames:
-            frame_gray = cv2.cvtColor(crop(frame, region), cv2.COLOR_BGR2GRAY)
-
-            imglog = logging.ImageLogger("detect_motion")
-            imglog.imwrite("source", frame_gray)
-
-            absdiff = cv2.absdiff(frame_gray, previous_frame_gray)
-            previous_frame_gray = frame_gray
-            imglog.imwrite("absdiff", absdiff)
-
-            if mask.image is not None:
-                absdiff = cv2.bitwise_and(absdiff, mask.image)
-                imglog.imwrite("mask", mask.image)
-                imglog.imwrite("absdiff_masked", absdiff)
-
-            _, thresholded = cv2.threshold(
-                absdiff, int((1 - noise_threshold) * 255), 255,
-                cv2.THRESH_BINARY)
-            eroded = cv2.erode(
-                thresholded,
-                cv2.getStructuringElement(cv2.MORPH_ELLIPSE, (3, 3)))
-            imglog.imwrite("absdiff_threshold", thresholded)
-            imglog.imwrite("absdiff_threshold_erode", eroded)
-
-            out_region = _pixel_bounding_box(eroded)
-            if out_region:
-                # Undo cv2.erode above:
-                out_region = out_region.extend(x=-1, y=-1)
-                # Undo crop:
-                out_region = out_region.translate(region.x, region.y)
-
-            motion = bool(out_region)
-
-            result = MotionResult(getattr(frame, "time", None), motion,
-                                  out_region, frame)
-            draw_on(frame, result, label="detect_motion()")
-            debug("%s found: %s" % (
-                "Motion" if motion else "No motion", str(result)))
-            yield result
-
     def wait_for_match(self, image, timeout_secs=10, consecutive_matches=1,
                        match_parameters=None, region=Region.ALL):
 
@@ -824,55 +669,6 @@ class DeviceUnderTest(object):
                     i += 1
                 else:
                     raise
-
-    def wait_for_motion(
-            self, timeout_secs=10, consecutive_frames=None,
-            noise_threshold=None, mask=None, region=Region.ALL):
-
-        if consecutive_frames is None:
-            consecutive_frames = get_config('motion', 'consecutive_frames')
-
-        consecutive_frames = str(consecutive_frames)
-        if '/' in consecutive_frames:
-            motion_frames = int(consecutive_frames.split('/')[0])
-            considered_frames = int(consecutive_frames.split('/')[1])
-        else:
-            motion_frames = int(consecutive_frames)
-            considered_frames = int(consecutive_frames)
-
-        if motion_frames > considered_frames:
-            raise ConfigurationError(
-                "`motion_frames` exceeds `considered_frames`")
-
-        debug("Waiting for %d out of %d frames with motion" % (
-            motion_frames, considered_frames))
-
-        if mask is None:
-            mask = _ImageFromUser(None, None, None)
-        else:
-            mask = _load_image(mask, cv2.IMREAD_GRAYSCALE)
-            debug("Using mask %s" % mask.friendly_name)
-
-        matches = deque(maxlen=considered_frames)
-        motion_count = 0
-        for res in self.detect_motion(timeout_secs, noise_threshold, mask,
-                                      region):
-            motion_count += bool(res)
-            if len(matches) == matches.maxlen:
-                motion_count -= bool(matches.popleft())
-            matches.append(res)
-            if motion_count >= motion_frames:
-                debug("Motion detected.")
-                # We want to return the first True motion result as this is when
-                # the motion actually started.
-                for result in matches:
-                    if result:
-                        return result
-                assert False, ("Logic error in wait_for_motion: This code "
-                               "should never be reached")
-
-        screenshot = self.get_frame()
-        raise MotionTimeout(screenshot, mask.friendly_name, timeout_secs)
 
     def ocr(self, frame=None, region=Region.ALL,
             mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD,
@@ -1276,30 +1072,6 @@ class MatchTimeout(UITestFailure):
     def __str__(self):
         return "Didn't find match for '%s' within %g seconds." % (
             self.expected, self.timeout_secs)
-
-
-class MotionTimeout(UITestFailure):
-    """Exception raised by `wait_for_motion`.
-
-    :ivar Frame screenshot: The last video frame that `wait_for_motion` checked
-        before timing out.
-
-    :vartype mask: str or None
-    :ivar mask: Filename of the mask that was used, if any.
-
-    :vartype timeout_secs: int or float
-    :ivar timeout_secs: Number of seconds that motion was searched for.
-    """
-    def __init__(self, screenshot, mask, timeout_secs):
-        super(MotionTimeout, self).__init__()
-        self.screenshot = screenshot
-        self.mask = mask
-        self.timeout_secs = timeout_secs
-
-    def __str__(self):
-        return "Didn't find motion%s within %g seconds." % (
-            " (with mask '%s')" % self.mask if self.mask else "",
-            self.timeout_secs)
 
 
 class PreconditionError(UITestError):
