+++ tests/test_motion.py
@@ -1,6 +1,8 @@
+import time
 from contextlib import contextmanager
 
 import numpy
+import pytest
 
 import stbt
 
@@ -18,51 +20,48 @@ def test_motionresult_repr():
 
 
 def test_wait_for_motion_half_motion_str_2of4():
-    with _fake_frames_at_half_motion() as dut:
-        res = dut.wait_for_motion(consecutive_frames='2/4')
-        print res
-        assert res.time == 1466084606.
+    with MockTime().patch():
+        res = stbt.wait_for_motion(
+            consecutive_frames='2/4', frames=fake_frames())
+    print res
+    assert res.time == 1466084606.
 
 
 def test_wait_for_motion_half_motion_str_2of3():
-    with _fake_frames_at_half_motion() as dut:
-        res = dut.wait_for_motion(consecutive_frames='2/3')
-        print res
-        assert res.time == 1466084606.
+    with MockTime().patch():
+        res = stbt.wait_for_motion(
+            consecutive_frames='2/3', frames=fake_frames())
+    print res
+    assert res.time == 1466084606.
 
 
 def test_wait_for_motion_half_motion_str_4of10():
-    with _fake_frames_at_half_motion() as dut:
+    with MockTime().patch():
         # Time is not affected by consecutive_frames parameter
-        res = dut.wait_for_motion(consecutive_frames='4/10', timeout_secs=20)
-        assert res.time == 1466084606.
+        res = stbt.wait_for_motion(
+            consecutive_frames='4/10', timeout_secs=20, frames=fake_frames())
+    assert res.time == 1466084606.
 
 
 def test_wait_for_motion_half_motion_str_3of4():
-    with _fake_frames_at_half_motion() as dut:
-        try:
-            dut.wait_for_motion(consecutive_frames='3/4')
-            assert False, "wait_for_motion succeeded unexpectedly"
-        except stbt.MotionTimeout:
-            pass
+    try:
+        with MockTime().patch():
+            stbt.wait_for_motion(consecutive_frames='3/4', frames=fake_frames())
+        assert False, "wait_for_motion succeeded unexpectedly"
+    except stbt.MotionTimeout:
+        pass
 
 
 def test_wait_for_motion_half_motion_int():
-    with _fake_frames_at_half_motion() as dut:
-        try:
-            dut.wait_for_motion(consecutive_frames=2)
-            assert False, "wait_for_motion succeeded unexpectedly"
-        except stbt.MotionTimeout:
-            pass
+    with pytest.raises(stbt.MotionTimeout), MockTime().patch():
+        stbt.wait_for_motion(consecutive_frames=2, frames=fake_frames())
 
 
-@contextmanager
-def _fake_frames_at_half_motion():
-    from _stbt.core import DeviceUnderTest, NoSinkPipeline
-
-    FRAMES = []
+def fake_frames():
     a = numpy.zeros((2, 2, 3), dtype=numpy.uint8)
+    a.flags.writeable = False
     b = numpy.ones((2, 2, 3), dtype=numpy.uint8) * 255
+    b.flags.writeable = False
 
     # Motion:                 v     v     v     v     v     v     v     v     v
     data = [a, a, a, a, a, a, b, b, a, a, b, b, a, a, b, b, a, a, b, b, a, a, b]
@@ -70,27 +69,47 @@ def _fake_frames_at_half_motion():
     #       |                 L Motion starts here at timestamp 1466084606.
     #       L Video starts here at timestamp 1466084600
 
-    FRAMES = [stbt.Frame(x, time=1466084600. + n) for n, x in enumerate(data)]
+    start_time = time.time()
+    for n, x in enumerate(data):
+        t = start_time + n
+        time.sleep(t - time.time())
+        yield stbt.Frame(x, time=t), t * 1e9
 
-    class FakeDisplay(object):
-        def get_frame(self, timeout_secs=10, since=0):  # pylint: disable=unused-argument
-            for f in FRAMES:
-                if f.time > since:
-                    return f
-            f = FRAMES[-1].copy()
-            f.time = since + 1
-            return f
 
-    class FakeTime(object):
-        def __init__(self, now):
-            self._now = now
+class MockTime(object):
+    def __init__(self, start_time=1466084600.):
+        self._time = start_time
+        self._functions = []
 
-        def time(self):
-            return self._now
+    def time(self):
+        t = self._time
+        return t
 
-        def sleep(self, duration):
-            self._now += duration
+    def sleep(self, seconds):
+        while self._functions and self._functions[0][0] <= self._time + seconds:
+            _, fn = self._functions.pop(0)
+            fn()
 
-    dut = DeviceUnderTest(display=FakeDisplay(), sink_pipeline=NoSinkPipeline(),
-                          _time=FakeTime(FRAMES[0].time))
-    yield dut
+        self._time += seconds
+
+    def interrupt(self, exception):
+        def raise_exception():
+            raise exception
+        self.at(0, raise_exception)
+
+    def at(self, offset, function):
+        self._functions.append((self._time + offset, function))
+        self._functions.sort()
+
+    @contextmanager
+    def assert_duration(self, seconds):
+        start_time = self._time
+        yield self
+        assert self._time - start_time == seconds
+
+    @contextmanager
+    def patch(self):
+        from mock import patch
+        with patch("time.time", self.time), \
+                patch("time.sleep", self.sleep):
+            yield self
